test_client/pyproject.toml
[project]
name = "test-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "smithery (>=0.1.1,<0.2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mcp (>=1.6.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "test_client", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
